fix: support maxFiles without slicing
We currently exceed `maxFiles` and slice the result later. This doesn't
work in an iterator since there is no end "result".

This change instead makes the walker stop walking when `maxFiles` is
met, so we already have the correctly sized result set.
