+++ src/api/functions/push-directory.ts
@@ -1,25 +1,28 @@
-import { FilterPredicate, Options } from "../../types";
+import { FilterPredicate, Options, Counts } from "../../types";
 
 export type PushDirectoryFunction = (
   directoryPath: string,
   paths: string[],
   pushPath: (path: string, arr: string[]) => void,
+  counts: Counts,
   filters?: FilterPredicate[]
 ) => void;
 
 function pushDirectoryWithRelativePath(root: string): PushDirectoryFunction {
-  return function (directoryPath, paths, pushPath) {
+  return function (directoryPath, paths, pushPath, counts) {
     pushPath(directoryPath.substring(root.length) || ".", paths);
+    counts.directories++;
   };
 }
 
 function pushDirectoryFilterWithRelativePath(
   root: string
 ): PushDirectoryFunction {
-  return function (directoryPath, paths, pushPath, filters) {
+  return function (directoryPath, paths, pushPath, counts, filters) {
     const relativePath = directoryPath.substring(root.length) || ".";
     if (filters!.every((filter) => filter(relativePath, true))) {
       pushPath(relativePath, paths);
+      counts.directories++;
     }
   };
 }
@@ -27,20 +30,24 @@ function pushDirectoryFilterWithRelativePath(
 const pushDirectory: PushDirectoryFunction = (
   directoryPath,
   paths,
-  pushPath
+  pushPath,
+  counts
 ) => {
   pushPath(directoryPath || ".", paths);
+  counts.directories++;
 };
 
 const pushDirectoryFilter: PushDirectoryFunction = (
   directoryPath,
   paths,
   pushPath,
+  counts,
   filters
 ) => {
   const path = directoryPath || ".";
   if (filters!.every((filter) => filter(path, true))) {
     pushPath(path, paths);
+    counts.directories++;
   }
 };
